Rearrange application folders for routing for faq page
