+++ src/app/app.module.ts
@@ -6,16 +6,34 @@ import { HttpClientModule } from '@angular/common/http';
 
 import { AppComponent } from './app.component';
 import { UnitsPipe } from './units.pipe';
+import {RouterModule, Routes} from '@angular/router';
+import { FaqComponent } from './faq/faq.component';
+import { HomeComponent } from './home/home.component';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: HomeComponent
+  },
+  {
+    path: 'faq',
+    component: FaqComponent
+  }
+];
+
 
 @NgModule({
   declarations: [
     AppComponent,
+    HomeComponent,
+    FaqComponent,
     UnitsPipe
   ],
   imports: [
     BrowserModule,
     FormsModule,
-    HttpClientModule
+    HttpClientModule,
+    RouterModule.forRoot(routes)
   ],
   providers: [],
   bootstrap: [AppComponent]
