+++ src/app/faq/faq.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-faq',
+  templateUrl: './faq.component.html',
+  styleUrls: ['./faq.component.scss']
+})
+export class FaqComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
